BraintreeApi/src/androidTest/java/com/braintreepayments/api/models/PostalAddressTest.java
package com.braintreepayments.api.models;

import android.support.test.runner.AndroidJUnit4;
import android.test.suitebuilder.annotation.SmallTest;

import org.json.JSONException;
import org.junit.Test;
import org.junit.runner.RunWith;

import static junit.framework.Assert.assertEquals;

@RunWith(AndroidJUnit4.class)
public class PostalAddressTest {

    @Test(timeout = 1000)
    @SmallTest
    public void constructsCorrectly() throws JSONException {
        PostalAddress postalAddress =
                new PostalAddress("123 Fake St.", "Apt. 3", "Oakland", "CA", "94602", "US");

        assertEquals("123 Fake St.", postalAddress.getStreetAddress());
        assertEquals("Apt. 3", postalAddress.getExtendedAddress());
        assertEquals("Oakland", postalAddress.getLocality());
        assertEquals("CA", postalAddress.getRegion());
        assertEquals("94602", postalAddress.getPostalCode());
        assertEquals("US", postalAddress.getCountryCodeAlpha2());
    }
}
